[chore][pkg/stanza] Cleanup retain operator files
