+++ pkg/stanza/operator/transformer/retain/transformer_test.go
